seasonal/season_1_dec-jan_2024/draft/telegram-tech-sketches/1_database/requirements.txt
botspot>=0.2.19
calmlib>=1.0.11

aiogram>=3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
motor>=3.3.0 